Bug Correction On DIV ID line to get a good anchor
This is correcting a bug when a user click on edit link. Now it will bring the user on the good line not at the beginning.

Bonus for Theme like Oblyon and fixed menu, i add a class "anchor". This let add in the theme the necessary padding to make this work right and do not let the line be hidden by the menu.